moved wheels to module
