+++ include/microros.hpp
@@ -9,29 +9,24 @@
 #include <rmw_microros/rmw_microros.h>
 #include <sensor_msgs/msg/battery_state.h>
 #include <sensor_msgs/msg/imu.h>
-#include <sensor_msgs/msg/joint_state.h>
 #include <sensor_msgs/msg/range.h>
-#include <std_msgs/msg/float32_multi_array.h>
 #include <std_msgs/msg/u_int32_multi_array.h>
 #include <std_msgs/msg/bool.h>
 #include <microros_transport/mbed_serial_transport.hpp>
 #include <rclc_parameter/rclc_parameter.h>
 
 #include <leds.hpp>
+#include <wheels.hpp>
 
 constexpr const uint8_t UXR_CLIENT_DOMAIN_ID_TO_OVERRIDE_WITH_ENV = 255;
 
 constexpr const char *NODE_NAME = "rosbot_ros2_firmware";
 constexpr const char *IMU_TOPIC_NAME = "_imu/data_raw";
 constexpr const char *WHEELS_STATE_TOPIC_NAME = "_motors_response";
-constexpr const char *BATTERY_TOPIC_NAME = "battery";
 constexpr const char *WHEELS_COMMAND_TOPIC_NAME = "_motors_cmd";
+constexpr const char *BATTERY_TOPIC_NAME = "battery";
 constexpr const char *SERVOS_COMMAND_TOPIC_NAME = "cmd_ser";
 
-constexpr const char *FRONT_LEFT_MOTOR_NAME = "fl_wheel_joint";
-constexpr const char *FRONT_RIGHT_MOTOR_NAME = "fr_wheel_joint";
-constexpr const char *REAR_LEFT_MOTOR_NAME = "rl_wheel_joint";
-constexpr const char *REAR_RIGHT_MOTOR_NAME = "rr_wheel_joint";
 enum Buttons{
     button_left,
     button_right,
@@ -46,14 +41,6 @@ enum Ranges {
     RANGES_COUNT
 };
 
-enum Motors {
-    motor_right_rear,
-    motor_left_rear,
-    motor_right_front,
-    motor_left_front,
-    MOTORS_COUNT
-};
-
 enum Servos{
     servo0,
     servo1,
@@ -64,12 +51,6 @@ enum Servos{
     SERVOS_COUNT
 };
 
-enum MotorsState {
-    motor_state_position,
-    motor_state_velocity,
-    motor_state_effort,
-    MOTORS_STATE_COUNT
-};
 
 enum AgentStates {
     WAITING_AGENT,
@@ -116,7 +97,6 @@ bool init_parameters();
 void fill_wheels_state_msg(sensor_msgs__msg__JointState *msg);
 void fill_imu_msg(sensor_msgs__msg__Imu *msg);
 void fill_battery_msg(sensor_msgs__msg__BatteryState *msg);
-void fill_wheels_command_msg(std_msgs__msg__Float32MultiArray *msg);
 void fill_servos_command_msg(std_msgs__msg__UInt32MultiArray *msg);
 void fill_range_msg(sensor_msgs__msg__Range *msg, uint8_t id);
